+++ resources/libraries/rbfreeradius_helpers.rb
@@ -1,5 +1,4 @@
 module Rbfreeradius
   module Helpers
-
   end
 end