+++ CHANGELOG.md
@@ -1,5 +1,17 @@
-freeradius CHANGELOG
+cookbook-freeradius CHANGELOG
 ===============
+
+## 0.0.5
+
+  - Miguel Negron
+    - [9e19d79] lint resources
+    - [d506038] lint recipes
+    - [722c45c] lint providers
+    - [679071e] lint libraries
+    - [3e63952] lint attributes
+    - [2ff458b] Update README.md
+    - [b4ece9c] Update rpm.yml
+    - [cf14abe] Update metadata.rb
 
 0.0.1
 -----